fix #8; refactore binding to separate classes
