+++ source/DD4T.DI.Ninject/Bootstrap.cs
@@ -40,15 +40,12 @@ namespace DD4T.DI.Ninject
 
             var load = Assembly.LoadFile(file);
 
-            var provider = AppDomain.CurrentDomain.GetAssemblies().Where(ass => ass.FullName.StartsWith("DD4T.Providers")).FirstOrDefault();
-            var providerTypes = provider.GetTypes();
-            var pageprovider = providerTypes.Where(a => typeof(IPageProvider).IsAssignableFrom(a)).FirstOrDefault();
-            var cpProvider = providerTypes.Where(a => typeof(IComponentPresentationProvider).IsAssignableFrom(a)).FirstOrDefault();
-            var linkProvider = providerTypes.Where(a => typeof(ILinkProvider).IsAssignableFrom(a)).FirstOrDefault();
-            var commonServices = providerTypes.Where(a => typeof(IProvidersCommonServices).IsAssignableFrom(a)).FirstOrDefault();
-            var binaryProvider = providerTypes.Where(a => typeof(IBinaryProvider).IsAssignableFrom(a)).FirstOrDefault();
-            var componentProvider = providerTypes.Where(a => typeof(IComponentProvider).IsAssignableFrom(a)).FirstOrDefault();
-
+            kernel.BindProviders();
+            kernel.BindFactories();
+            kernel.BindMvc();
+            kernel.BindRestProvider();
+            kernel.BindResolvers();
+            kernel.BindViewModels();
 
             if (kernel.TryGet<IDD4TConfiguration>() == null)
                 kernel.Bind<IDD4TConfiguration>().To<DD4TConfiguration>().InSingletonScope();
@@ -56,69 +53,8 @@ namespace DD4T.DI.Ninject
             if (kernel.TryGet<ILogger>() == null)
                 kernel.Bind<ILogger>().To<DefaultLogger>().InSingletonScope();
 
-            if (kernel.TryGet<IPublicationResolver>() == null)
-                kernel.Bind<IPublicationResolver>().To<DefaultPublicationResolver>().InSingletonScope();
-
             if (kernel.TryGet<ICacheAgent>() == null)
                 kernel.Bind<ICacheAgent>().To<DefaultCacheAgent>();
-
-
-            kernel.BindMVCTypes();
-            kernel.BindRestProviderTypes();
-
-            //providers
-            if (binaryProvider != null && kernel.TryGet<IBinaryProvider>() == null)
-                kernel.Bind<IBinaryProvider>().To(binaryProvider);
-
-            if (componentProvider != null && kernel.TryGet<IComponentProvider>() == null)
-                kernel.Bind<IComponentProvider>().To(componentProvider);
-
-            if (pageprovider != null && kernel.TryGet<IPageProvider>() == null)
-                kernel.Bind<IPageProvider>().To(pageprovider);
-
-            if (cpProvider != null && kernel.TryGet<IComponentPresentationProvider>() == null)
-                kernel.Bind<IComponentPresentationProvider>().To(cpProvider);
-
-            if (linkProvider != null && kernel.TryGet<ILinkProvider>() == null)
-                kernel.Bind<ILinkProvider>().To(linkProvider);
-
-            if (commonServices != null && kernel.TryGet<IProvidersCommonServices>() == null)
-                kernel.Bind<IProvidersCommonServices>().To(commonServices);
-
-            //factories
-            if (kernel.TryGet<IPageFactory>() == null)
-                kernel.Bind<IPageFactory>().To<PageFactory>();
-
-            if (kernel.TryGet<IComponentPresentationFactory>() == null)
-                kernel.Bind<IComponentPresentationFactory>().To<ComponentPresentationFactory>();
-
-            if (kernel.TryGet<ILinkFactory>() == null)
-                kernel.Bind<ILinkFactory>().To<LinkFactory>();
-
-            if (kernel.TryGet<IBinaryFactory>() == null)
-                kernel.Bind<IBinaryFactory>().To<BinaryFactory>();
-
-            if(kernel.TryGet<IComponentFactory>() ==  null)
-                kernel.Bind<IComponentFactory>().To<ComponentFactory>();
-
-            if (kernel.TryGet<IFactoryCommonServices>() == null)
-                kernel.Bind<IFactoryCommonServices>().To<FactoryCommonServices>();
-
-
-
-            //factories
-            if (kernel.TryGet<IReflectionHelper>() == null)
-                kernel.Bind<IReflectionHelper>().To<ReflectionOptimizer>().InSingletonScope();
-
-            if (kernel.TryGet<IViewModelResolver>() == null)
-                kernel.Bind<IViewModelResolver>().To<DefaultViewModelResolver>().InSingletonScope();
-
-            if (kernel.TryGet<IViewModelFactory>() == null)
-                kernel.Bind<IViewModelFactory>().To<ViewModelFactory>().InSingletonScope();
-
-            if (kernel.TryGet<IViewModelKeyProvider>() == null)
-                kernel.Bind<IViewModelKeyProvider>().To<WebConfigViewModelKeyProvider>().InSingletonScope();
-
 
             //caching JMS
             if (kernel.TryGet<IMessageProvider>() == null)
