+++ source/DD4T.DI.Ninject/DD4T.DI.Ninject.csproj
@@ -33,36 +33,36 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="DD4T.ContentModel, Version=2.0.4.0, Culture=neutral, PublicKeyToken=4450e3c7f68bf872, processorArchitecture=MSIL">
-      <HintPath>..\packages\DD4T.Model.2.0.4-beta1\lib\net45\DD4T.ContentModel.dll</HintPath>
+    <Reference Include="DD4T.ContentModel">
+      <HintPath>..\packages\DD4T.Model.2.0.5.19-alpha\lib\net45\DD4T.ContentModel.dll</HintPath>
       <Private>False</Private>
     </Reference>
-    <Reference Include="DD4T.ContentModel.Contracts, Version=2.0.4.0, Culture=neutral, PublicKeyToken=4450e3c7f68bf872, processorArchitecture=MSIL">
-      <HintPath>..\packages\DD4T.Model.2.0.4-beta1\lib\net45\DD4T.ContentModel.Contracts.dll</HintPath>
+    <Reference Include="DD4T.ContentModel.Contracts">
+      <HintPath>..\packages\DD4T.Model.2.0.5.19-alpha\lib\net45\DD4T.ContentModel.Contracts.dll</HintPath>
       <Private>False</Private>
     </Reference>
-    <Reference Include="DD4T.ContentModel.XmlSerializers, Version=2.0.4.0, Culture=neutral, PublicKeyToken=4450e3c7f68bf872, processorArchitecture=MSIL">
-      <HintPath>..\packages\DD4T.Model.2.0.4-beta1\lib\net45\DD4T.ContentModel.XmlSerializers.dll</HintPath>
+    <Reference Include="DD4T.ContentModel.XmlSerializers">
+      <HintPath>..\packages\DD4T.Model.2.0.5.19-alpha\lib\net45\DD4T.ContentModel.XmlSerializers.dll</HintPath>
       <Private>False</Private>
     </Reference>
-    <Reference Include="DD4T.Core.Contracts, Version=2.0.5.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\DD4T.Core.2.0.5-beta\lib\net45\DD4T.Core.Contracts.dll</HintPath>
+    <Reference Include="DD4T.Core.Contracts">
+      <HintPath>..\packages\DD4T.Core.2.0.6.95-alpha\lib\net45\DD4T.Core.Contracts.dll</HintPath>
       <Private>False</Private>
     </Reference>
-    <Reference Include="DD4T.Factories, Version=2.0.5.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\DD4T.Core.2.0.5-beta\lib\net45\DD4T.Factories.dll</HintPath>
+    <Reference Include="DD4T.Factories">
+      <HintPath>..\packages\DD4T.Core.2.0.6.95-alpha\lib\net45\DD4T.Factories.dll</HintPath>
       <Private>False</Private>
     </Reference>
-    <Reference Include="DD4T.Serialization, Version=2.0.4.0, Culture=neutral, PublicKeyToken=4450e3c7f68bf872, processorArchitecture=MSIL">
-      <HintPath>..\packages\DD4T.Model.2.0.4-beta1\lib\net45\DD4T.Serialization.dll</HintPath>
+    <Reference Include="DD4T.Serialization">
+      <HintPath>..\packages\DD4T.Model.2.0.5.19-alpha\lib\net45\DD4T.Serialization.dll</HintPath>
       <Private>False</Private>
     </Reference>
-    <Reference Include="DD4T.Utils, Version=2.0.5.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\DD4T.Core.2.0.5-beta\lib\net45\DD4T.Utils.dll</HintPath>
+    <Reference Include="DD4T.Utils">
+      <HintPath>..\packages\DD4T.Core.2.0.6.95-alpha\lib\net45\DD4T.Utils.dll</HintPath>
       <Private>False</Private>
     </Reference>
-    <Reference Include="DD4T.ViewModels, Version=2.0.5.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\DD4T.Core.2.0.5-beta\lib\net45\DD4T.ViewModels.dll</HintPath>
+    <Reference Include="DD4T.ViewModels">
+      <HintPath>..\packages\DD4T.Core.2.0.6.95-alpha\lib\net45\DD4T.ViewModels.dll</HintPath>
       <Private>False</Private>
     </Reference>
     <Reference Include="Ninject, Version=3.2.0.0, Culture=neutral, PublicKeyToken=c7192dc5380945e7, processorArchitecture=MSIL">
@@ -80,9 +80,13 @@
   <ItemGroup>
     <Compile Include="Bootstrap.cs" />
     <Compile Include="Exceptions\ProviderNotFoundException.cs" />
-    <Compile Include="MVCTypes.cs" />
+    <Compile Include="Factories.cs" />
+    <Compile Include="Mvc.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RestProviderTypes.cs" />
+    <Compile Include="Providers.cs" />
+    <Compile Include="Resolvers.cs" />
+    <Compile Include="RestProvider.cs" />
+    <Compile Include="ViewModels.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
