+++ source/DD4T.DI.Ninject/packages.config
@@ -1,7 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="DD4T.Core" version="2.0.5-beta" targetFramework="net451" />
-  <package id="DD4T.Model" version="2.0.4-beta1" targetFramework="net451" />
+  <package id="DD4T.Core" version="2.0.6.95-alpha" targetFramework="net451" />
+  <package id="DD4T.Model" version="2.0.5.19-alpha" targetFramework="net451" />
   <package id="Newtonsoft.Json" version="6.0.8" targetFramework="net451" />
   <package id="Ninject" version="3.2.2.0" targetFramework="net451" />
 </packages>
